+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for YAML configurations.
 - Allow specifying external config/rdf files.
 - Added support for trig files.
+- Import vocabularies every hour using cron.
 
 ## [v2.15-1.1.0]