+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
 - Added support for YAML configurations.
 - Allow specifying external config/rdf files.
 - Added support for trig files.
 - Import vocabularies every hour using cron.
+- Added a configuration option for the refresh interval of vocabularies.
 
 ## [v2.15-1.1.0]